+++ projects/js-packages/charts/src/providers/theme/themes.ts
@@ -8,12 +8,14 @@ const defaultTheme: ChartTheme = {
 	labelBackgroundColor: '#FFFFFF', // label background color
 	colors: [ '#98C8DF', '#006DAB', '#A6DC80', '#1F9828', '#FF8C8F' ],
 	gridStyles: {
-		stroke: '#787C82',
+		stroke: '#DCDCDE',
 		strokeWidth: 1,
 	},
-	tickLength: 0,
+	tickLength: 4,
 	gridColor: '',
 	gridColorDark: '',
+	xTickLineStyles: { stroke: 'black' },
+	xAxisLineStyles: { stroke: '#DCDCDE', strokeWidth: 1 },
 };
 
 /**
@@ -24,12 +26,14 @@ const jetpackTheme: ChartTheme = {
 	labelBackgroundColor: '#FFFFFF', // label background color
 	colors: [ '#98C8DF', '#006DAB', '#A6DC80', '#1F9828', '#FF8C8F' ],
 	gridStyles: {
-		stroke: '#787C82',
+		stroke: '#DCDCDE',
 		strokeWidth: 1,
 	},
-	tickLength: 0,
+	tickLength: 4,
 	gridColor: '',
 	gridColorDark: '',
+	xTickLineStyles: { stroke: 'black' },
+	xAxisLineStyles: { stroke: '#DCDCDE', strokeWidth: 1 },
 };
 
 /**
@@ -43,9 +47,11 @@ const wooTheme: ChartTheme = {
 		stroke: '#787C82',
 		strokeWidth: 1,
 	},
-	tickLength: 0,
+	tickLength: 4,
 	gridColor: '',
 	gridColorDark: '',
+	xTickLineStyles: { stroke: 'black' },
+	xAxisLineStyles: { stroke: '#DCDCDE', strokeWidth: 1 },
 };
 
 export { defaultTheme, jetpackTheme, wooTheme };