+++ projects/packages/forms/src/blocks/contact-form/class-contact-form-block.php
@@ -30,6 +30,44 @@ class Contact_Form_Block {
 				'render_callback' => array( __CLASS__, 'gutenblock_render_form' ),
 			)
 		);
+
+		add_filter( 'render_block_data', array( __CLASS__, 'find_nested_html_block' ), 10, 3 );
+		add_filter( 'render_block_core/html', array( __CLASS__, 'render_wrapped_html_block' ), 10, 2 );
+	}
+
+	/**
+	 *  Find nested html block that reside in the contact form block.
+	 *  We are using this to wrap the html block with div if it is nested inside contact form block. So that the elements render as expected.
+	 *
+	 *  @param array  $parsed_block - the parsed block.
+	 *  @param array  $source_block - the source block.
+	 *  @param object $parent_block - the parent WP_Block.
+	 *
+	 *  @return array
+	 */
+	public static function find_nested_html_block( $parsed_block, $source_block, $parent_block ) {
+		if ( $parsed_block['blockName'] === 'core/html' && isset( $parent_block->parsed_block ) && $parent_block->parsed_block['blockName'] === 'jetpack/contact-form' ) {
+			$parsed_block['hasJPFormParent'] = true;
+		}
+		return $parsed_block;
+	}
+
+	/**
+	 * Render wrapped html block that is inside the form block with a wrapped div so that the elements render as expected.
+	 * The extra div is needed because the form block has a `flex: 0 0 100%;` applied to all the children of the form block.
+	 * This cases all the elementes inside the block to render in a single line and make it not possible to add have inline elements.
+	 *
+	 * @param string $content - the content of the block.
+	 * @param array  $parsed_block - the parsed block.
+	 *
+	 * @return string
+	 */
+	public static function render_wrapped_html_block( $content, $parsed_block ) {
+		if ( ! empty( $parsed_block['hasJPFormParent'] ) ) {
+			return '<div>' . $content . '</div>';
+		}
+
+		return $content;
 	}
 
 	/**
@@ -156,7 +194,6 @@ class Contact_Form_Block {
 	 * Loads scripts
 	 */
 	public static function load_editor_scripts() {
-		global $post;
 
 		$handle = 'jp-forms-blocks';
 
@@ -171,10 +208,14 @@ class Contact_Form_Block {
 			)
 		);
 
+		// Create a Contact_Form instance to get the default values
+		$contact_form = new Contact_Form( array() );
+		$defaults     = $contact_form->defaults;
+
 		$data = array(
 			'defaults' => array(
-				'to'      => wp_get_current_user()->user_email,
-				'subject' => '[' . get_bloginfo( 'name' ) . ']' . ( isset( $post ) ? ' ' . esc_html( $post->post_title ) : '' ),
+				'to'      => $defaults['to'],
+				'subject' => $defaults['subject'],
 			),
 		);