+++ projects/plugins/jetpack/extensions/plugins/ai-assistant-plugin/components/seo-assistant/use-keywords-step.tsx
@@ -1,36 +1,56 @@
-import { createInterpolateElement, useCallback, useState } from '@wordpress/element';
+import { createInterpolateElement, useCallback, useState, useEffect } from '@wordpress/element';
 import { __, sprintf } from '@wordpress/i18n';
-import type { Step } from './types';
+import TypingMessage from './typing-message';
+import { useMessages } from './wizard-messages';
+import type { InputStep } from './types';
 
-export const useKeywordsStep = ( { addMessage, onStep } ): Step => {
+export const useKeywordsStep = (): InputStep => {
 	const [ keywords, setKeywords ] = useState( '' );
 	const [ completed, setCompleted ] = useState( false );
+	const { messages, setMessages, addMessage, removeLastMessage } = useMessages();
+
+	useEffect( () => {
+		setMessages( [
+			{
+				content: __(
+					'To start, please enter 1–3 focus keywords that describe your blog post.',
+					'jetpack'
+				),
+				showIcon: true,
+			},
+		] );
+	}, [ setMessages ] );
 
 	const handleSkip = useCallback( () => {
 		addMessage( { content: __( 'Skipped!', 'jetpack' ) } );
-		if ( onStep ) {
-			onStep( { value: '' } );
-		}
-	}, [ addMessage, onStep ] );
+	}, [ addMessage ] );
 
-	const handleKeywordsSubmit = useCallback( () => {
+	const handleKeywordsSubmit = useCallback( async () => {
 		if ( ! keywords.trim() ) {
 			return handleSkip();
 		}
 		addMessage( { content: keywords, isUser: true } );
+		addMessage( { content: <TypingMessage /> } );
 
-		const keywordlist = keywords
-			.split( ',' )
-			.map( k => k.trim() )
-			.reduce( ( acc, curr, i, arr ) => {
-				if ( arr.length === 1 ) {
-					return curr;
-				}
-				if ( i === arr.length - 1 ) {
-					return `${ acc } </b>&<b> ${ curr }`;
-				}
-				return i === 0 ? curr : `${ acc }, ${ curr }`;
-			}, '' );
+		const keywordlist = await new Promise( resolve =>
+			setTimeout( () => {
+				const commaSeparatedKeywords = keywords
+					.split( ',' )
+					.map( k => k.trim() )
+					.reduce( ( acc, curr, i, arr ) => {
+						if ( arr.length === 1 ) {
+							return curr;
+						}
+						if ( i === arr.length - 1 ) {
+							return `${ acc } </b>&<b> ${ curr }`;
+						}
+						return i === 0 ? curr : `${ acc }, ${ curr }`;
+					}, '' );
+				resolve( commaSeparatedKeywords );
+			}, 500 )
+		);
+		removeLastMessage();
+
 		const message = createInterpolateElement(
 			/* Translators: wrapped string is list of keywords user has entered */
 			sprintf( __( `Got it! You're targeting <b>%s</b>. ✨✅`, 'jetpack' ), keywordlist ),
@@ -40,41 +60,13 @@ export const useKeywordsStep = ( { addMessage, onStep } ): Step => {
 		);
 		addMessage( { content: message } );
 		setCompleted( true );
-		if ( onStep ) {
-			onStep( { value: keywords } );
-		}
-	}, [ onStep, addMessage, keywords, handleSkip ] );
+	}, [ addMessage, keywords, handleSkip, removeLastMessage ] );
 
 	return {
 		id: 'keywords',
 		title: __( 'Optimise for SEO', 'jetpack' ),
 		label: __( 'Keywords', 'jetpack' ),
-		messages: [
-			{
-				content: createInterpolateElement(
-					__( "<b>Hi there! 👋 Let's optimise your blog post for SEO.</b>", 'jetpack' ),
-					{ b: <b /> }
-				),
-				showIcon: true,
-			},
-			{
-				content: createInterpolateElement(
-					__(
-						"Here's what we can improve:<br />1. Keywords<br />2. Title<br />3. Meta description",
-						'jetpack'
-					),
-					{ br: <br /> }
-				),
-				showIcon: false,
-			},
-			{
-				content: __(
-					'To start, please enter 1–3 focus keywords that describe your blog post.',
-					'jetpack'
-				),
-				showIcon: true,
-			},
-		],
+		messages,
 		type: 'input',
 		placeholder: __( 'Photography, plants', 'jetpack' ),
 		onSubmit: handleKeywordsSubmit,