+++ projects/packages/videopress/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.25.8] - 2025-01-27
+### Fixed
+- VideoPress: Fix issue with VideoPress block with 0 height and width [#41319]
+
 ## [0.25.7] - 2025-01-20
 ### Changed
 - Code: Use function-style exit() and die() with a default status code of 0. [#41167]
@@ -1517,6 +1521,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Created empty package [#24952]
 
+[0.25.8]: https://github.com/Automattic/jetpack-videopress/compare/v0.25.7...v0.25.8
 [0.25.7]: https://github.com/Automattic/jetpack-videopress/compare/v0.25.6...v0.25.7
 [0.25.6]: https://github.com/Automattic/jetpack-videopress/compare/v0.25.5...v0.25.6
 [0.25.5]: https://github.com/Automattic/jetpack-videopress/compare/v0.25.4...v0.25.5